Version 1.1: Changes to pricing tree generation
Updated method used to generate the pricing tree. Changes have led to an extreme increase in efficiency, script can now perform a 1000 branch valuation in under 0.5 seconds and a 10000 branch valuation in under 50 seconds, with printing off. The removal of list sorting and item deletion has reduced demand on memory when utilising a large number of branches.